Added animation to notification overlay. Refactored notifications folder.
