--- src/components/NotificationPopover/index.js
@@ -1,1 +1,0 @@
-export * from './component';
