+++ src/screens/Notifications/NotificationHeader/component.js
