+++ src/screens/Notifications/NotificationHeader/index.js
@@ -1,0 +1,1 @@
+export * from './component';
