+++ src/screens/Notifications/NotificationOverlay/components/NotificationPopover/component.js
@@ -5,15 +5,15 @@ import { PropTypes } from 'prop-types';
 import { parseNotificationData } from 'kitsu/utils/notifications';
 import { styles } from './styles';
 
-export const NotificationPopover = ({ style, onRequestClose, data }) => {
+export const NotificationPopover = ({ style, onPress, data }) => {
   if (!data) return null;
   const { actorName, actorAvatar, others, text } = parseNotificationData(data.activities);
 
   return (
     <View style={style}>
-      <TouchableOpacity activeOpacity={1} onPress={onRequestClose} style={styles.wrapper}>
+      <TouchableOpacity activeOpacity={1} onPress={onPress} style={styles.wrapper}>
         <StatusBar translucent backgroundColor={'rgba(0, 0, 0, 0.3)'} barStyle={'light-content'} />
-        <TouchableOpacity activeOpacity={0.9} onPress={onRequestClose} style={styles.content}>
+        <TouchableOpacity activeOpacity={0.9} onPress={onPress} style={styles.content}>
           <View style={{ paddingRight: 10 }}>
             <FastImage style={styles.userAvatar} source={{ uri: actorAvatar }} cache="web" />
           </View>
@@ -33,7 +33,7 @@ export const NotificationPopover = ({ style, onRequestClose, data }) => {
 NotificationPopover.propTypes = {
   style: ViewPropTypes.style,
   data: PropTypes.object,
-  onRequestClose: PropTypes.func.isRequired,
+  onPress: PropTypes.func.isRequired,
 };
 NotificationPopover.defaultProps = {
   style: null,
